requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
python-multipart==0.0.9

